requirements.txt
django
django-extensions
python-decouple
django-bootstrap5
whitenoise
django-ckeditor
Pillow